Move Stripe webhook logic into dedicated frontend controller
