+++ app/controllers/admin/stripe_accounts_controller.rb
@@ -42,19 +42,6 @@ module Admin
       redirect_to spree.admin_path
     end
 
-    def deauthorize
-      # TODO is there a sensible way to confirm this webhook call is actually from Stripe?
-      event = Stripe::Event.construct_from(params)
-      return render nothing: true, status: 204 unless event.type == "account.application.deauthorized"
-
-      destroyed = StripeAccount.where(stripe_user_id: event.account).destroy_all
-      if destroyed.any?
-        render text: "Account #{event.account} deauthorized", status: 200
-      else
-        render nothing: true, status: 204
-      end
-    end
-
     def status
       return render json: { status: :stripe_disabled } unless Spree::Config.stripe_connect_enabled
       stripe_account = StripeAccount.find_by_enterprise_id(params[:enterprise_id])
